fix/issue-107: Prevent lucene corrupting vault index when switching locations.
Create a separate .searchIndex lucene folder local in the vault to prevent different vaults using the same search index.

Stop the indexer before changing the vault location.
It is to make sure the indexer does not write data from the current vault to the new location.

Replace deprecated code that saves the settings file with standard Java library method.