feature/search-improvements Features to improve linking and searching.
Automatically assign a unique identifier (Uuid) to each document and save
it in the meta data. This allows you to refer to the document even if the
title or the content changes. Added the uuid to the document header bar,
you can select and copy the identifier.

Added a new search prefix uuid: with which the user can search for
the document having that identifier. In markdown links of the form
"[My Other Doc](search://uuid:42c73fbe-16c5-4b46-8477-6bb0a36d73d1)
are possible. To make the process easier a menu item "Copy Markdown Link"
was added that will copy a link to the clipboard which can be pasted in
another Markdown document.

Modified the semantics of title: tag: and nb: searches. Until now,
only the first word following the title prefix was used for searching.
It was impossible to search for notebooks, notes or tags containing
spaces, and the spaces were not indexed anyway.
Improvements:
 * The indexer now addes the full title, tag and notebook name
   (spaces included) to the index (in addition to the separate words
   in the title).
 * Added double quote delimiters for strings eg. "my document title" so
   that you can search for entire titles, tags and notebook names
   using the complete strings, eg. title:"my document title"
 * Added %22 code to html to allow the delimited strings in search urls.
   You can link to another document using its title
   [my other doc](search://title:%22my+other+doc%22).
   In most cases it is better to link using the uuid because it is
   independent of name changes, but if you want it you can link by title.